Move outdated logic to legacy folder
This moves the logic to extract SWE data based on shapefiles into
a separate folder. Probably not needed at all and preserving for now
as reference on what was tried.
