+++ legacy/rasterize_zone.py
